refactor(game):css파일명 변경에따른 경로 수정
